# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: EffiPlat 全系统 API 文档
  version: "1.0"
  description: |
    EffiPlat 平台后端 API，涵盖用户、角色、权限、职责、环境、资产、服务、业务、Bug、认证等核心模块。
servers:
  - url: /api/v1
...
paths:
...
  /environments:
    get:
      tags: [environments]
      summary: 获取环境列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 10 }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: code
          schema: { type: string }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentListResponse'
    post:
      tags: [environments]
      summary: 创建环境
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentResponse'
  /environments/{envId}:
    get:
      tags: [environments]
      summary: 获取环境详情
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: envId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentDetailsResponse'
    put:
      tags: [environments]
      summary: 更新环境
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: envId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentResponse'
    delete:
      tags: [environments]
      summary: 删除环境
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: envId
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: 删除成功，无响应体

  # 资产管理
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Environment:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string, description: "URL-friendly identifier for the environment" }
        description: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    EnvironmentListResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string }
        data:
          type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Environment' }
            total: { type: integer }
            page: { type: integer }
            pageSize: { type: integer }
...
    EnvironmentCreateRequest:
      type: object
      required: [name, slug]
      properties:
        name: { type: string, minLength: 2, maxLength: 100, description: "Name of the environment" }
        description: { type: [string, "null"], description: "Description of the environment" }
        slug: { type: string, minLength: 2, maxLength: 50, pattern: "^[a-zA-Z0-9_-]+$", description: "URL-friendly slug (alphanumeric, dash, underscore)" }
    EnvironmentUpdateRequest:
      type: object
      properties:
        name: { type: [string, "null"], minLength: 2, maxLength: 100, description: "Name of the environment" }
        description: { type: [string, "null"], description: "Description of the environment" }
        slug: { type: [string, "null"], minLength: 2, maxLength: 50, pattern: "^[a-zA-Z0-9_-]+$", description: "URL-friendly slug (alphanumeric, dash, underscore)" }
    EnvironmentResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: "Success" }
        data:
          $ref: '#/components/schemas/Environment'
    EnvironmentDetailsResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: "Success" }
        data:
          $ref: '#/components/schemas/Environment'

    # Next: Check and adjust EnvironmentListResponse
...